refactor(juncture public key): refactored oauth controllers to take in the public key in a header instead + updated documentation to reflect
