+++ server/src/utils/juncture_key_helpers/public_key_helpers.ts
@@ -1,0 +1,10 @@
+import { Request } from "express";
+
+export async function extractPublicKeyFromRequest(req: Request): Promise<string | undefined> {
+    const publicKey = req.get("x-juncture-public-key");
+    if (!publicKey) {
+        return undefined;
+    }
+
+    return publicKey;
+}
